[FIX] sale_timesheet : hide remaining hours on subscription tasks
This change hides the remaining_hours_so field when the sales order line is linked to a subscription. Unlike standard service or time-based sales orders, where this field reflects the difference between the quantity ordered and the quantity delivered, the concept does not translate well to subscription logic.

In the context of a subscription, the service is delivered on a recurring period (monthly, yearly, etc.). Delivery quantities continuously accumulate over time, and because the subscription renews indefinitely until cancellation, the “remaining hours” calculation quickly becomes misleading. In many cases it can drift into negative values, giving the impression of an error or over-consumption when, in reality, the subscription is simply following its recurring delivery cycle.

To avoid confusing end-users and to maintain a clean, intuitive interface, we hide this field whenever the line is part of a subscription.

opw-5246238

closes odoo/odoo#247685

X-original-commit: 6a23b2939f2cfc8c3a550d1ece598d48676f84d2
Related: odoo/enterprise#106794
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>